# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for our application.",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/voucher_claims": {
            "get": {
                "summary": "Get list of voucher claims",
                "operationId": "f369ac7c8996f86a7102fab335ddd364",
                "responses": {
                    "200": {
                        "description": "List of voucher claims",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VoucherClaim"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new voucher claim",
                "operationId": "fc6e846cbe08c6ac97c393cd07b65ddd",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoucherClaim"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voucher claim created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoucherClaim"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/voucher_claims/{id}": {
            "get": {
                "summary": "Get a specific voucher claim",
                "operationId": "e1e1c3a25978b499220a736771fefe3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Voucher Claim ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voucher claim details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoucherClaim"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "summary": "Update a voucher claim",
                "operationId": "42f9be8a87af23b828c2f778ff47566f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Voucher Claim ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoucherClaim"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voucher claim updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoucherClaim"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a voucher claim",
                "operationId": "9a13032321884d40a03d63cfc5d22a85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Voucher Claim ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Voucher claim deleted"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/vouchers": {
            "get": {
                "summary": "Get list of vouchers",
                "operationId": "2246b623ce7a9d5bd2220612b816d781",
                "parameters": [
                    {
                        "name": "kategori",
                        "in": "query",
                        "description": "Kategori vouchers",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of vouchers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Voucher"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new voucher",
                "operationId": "0a474a047c834069873a53dddad25a9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Voucher"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Voucher created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Voucher"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/vouchers/{id}": {
            "get": {
                "summary": "Get a specific voucher",
                "operationId": "734100359714d9ba10b1e59978d488dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Voucher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Voucher details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Voucher"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "summary": "Update a voucher",
                "operationId": "a3539251e17d483bc69ad8a86197bca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Voucher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Voucher"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Voucher updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Voucher"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a voucher",
                "operationId": "173dfc59a6e61236e8e2512db33064c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Voucher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Voucher deleted"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/vouchers/restore/{id}": {
            "post": {
                "summary": "Restore a voucher",
                "operationId": "c15be2e4315a0d4e659a65977403d4ea",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Voucher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Voucher restored"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Voucher": {
                "required": [
                    "nama",
                    "kategori",
                    "status"
                ],
                "properties": {
                    "nama": {
                        "type": "string"
                    },
                    "foto": {
                        "type": "string"
                    },
                    "kategori": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "VoucherClaim": {
                "required": [
                    "id_voucher",
                    "tanggal_claim"
                ],
                "properties": {
                    "id_voucher": {
                        "type": "string"
                    },
                    "tanggal_claim": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Authentication Bearer Token",
                "name": "Authentication Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}